+++ src/main/java/cn/dsgrp/field/stock/web/account/UserAdminController.java
@@ -5,6 +5,7 @@
  *******************************************************************************/
 package cn.dsgrp.field.stock.web.account;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import javax.validation.Valid;
@@ -42,7 +43,7 @@ public class UserAdminController {
 	}
 
 	@RequestMapping(value = "update/{id}", method = RequestMethod.GET)
-	public String updateForm(@PathVariable("id") Long id, Model model) {
+	public String updateForm(@PathVariable("id") BigInteger id, Model model) {
 		model.addAttribute("user", accountService.getUser(id));
 		return "account/adminUserForm";
 	}
@@ -55,7 +56,7 @@ public class UserAdminController {
 	}
 
 	@RequestMapping(value = "delete/{id}")
-	public String delete(@PathVariable("id") Long id, RedirectAttributes redirectAttributes) {
+	public String delete(@PathVariable("id") BigInteger id, RedirectAttributes redirectAttributes) {
 		User user = accountService.getUser(id);
 		accountService.deleteUser(id);
 		redirectAttributes.addFlashAttribute("message", "删除用户" + user.getLoginName() + "成功");
@@ -67,8 +68,8 @@ public class UserAdminController {
 	 * 因为仅update()方法的form中有id属性，因此仅在update时实际执行.
 	 */
 	@ModelAttribute
-	public void getUser(@RequestParam(value = "id", defaultValue = "-1") Long id, Model model) {
-		if (id != -1) {
+	public void getUser(@RequestParam(value = "id", defaultValue = "-1") BigInteger id, Model model) {
+		if (id.longValue() != -1) {
 			model.addAttribute("user", accountService.getUser(id));
 		}
 	}
